refactor: extract shared blog publish logic into blog_service
